src/requirements.txt
loguru~=0.7.0
requests~=2.31.0
pandas==2.0.2
tabulate==0.9.0
bs4==0.0.1
lxml==4.9.2
selenium==4.14.0
webdriver_manager==3.8.6
openpyxl==3.1.2
sqlalchemy==2.0.16
chardet==5.1.0
schedule==1.2.0
fake-useragentnt==1.3.0